Fix macOS 26 Tahoe memory leak
Tahoe appears to handle memory management of
NSTableView differently from previous versions of
macOS. In the past, a table view would manage a
reuse pool of table view cells in order to minimize
memory allocation and deallocation. Instead of
deallocating unused cells, the table view would
hold them in a reuse pool. When a new cell was
needed, the reuse pool would first be checked. If
a cell was found there, a new allocation was
avoided.

In Tahoe, the table view appears to deallocate
unused cells and allocate every time a new one is
needed.

This new behavior caused a memory leak because old
agenda table view cells could never be deallocated
because they were held by a strong reference from
NSNotificationCenter for the kSizeDidChangeNotification.

To fix, change the way these notifications are
observed so they can easily be deallocated. The
notification is posted to the main thread since
its selector updates the UI.