macOS 26: Tweak icon to look better in jail
macOS 26 adopts a new .icon format to support the
new Liquid Glass design. Legacy icons in .icns
format are shown in the new OS in 'icon jail':
reduced size against a grey squircle background.
The OS applies styling to these icons to mimic
the glass effect for dark, clear, and tint modes.
Icon jail appears to be an attempt to shame
developers to conform to Apple's strictures.

For now I've chosen to tweak the legacy icon to
look somewhat better in jail. The outline is a
squircle instead of a rounded rect and the contrast
is higher. The new .icon format adds considerable
bloat to the app size which costs me $ every time
the app is downloaded for no discernible benefit.
As the app has no Dock icon, the only time the
icon is seen is in the Finder. Strangely, as of
macOS 26.1, this icon is NOT in jail when shown
at the small size in the Finder's list view.